+++ server/getObject_test.go
@@ -222,6 +222,7 @@ func setupFakeServerWithObjectForUser(user models.ODUser, obj models.ODObject) *
 		SnippetResp:        &snippetResponse,
 		UserAttributesResp: &attributesResponse,
 	}
+
 	fakeQueue := kafka.NewFakeAsyncProducer(nil)
 	fakeServer := server.AppServer{RootDAO: &fakeDAO,
 		AAC:             &fakeAAC,
@@ -229,6 +230,7 @@ func setupFakeServerWithObjectForUser(user models.ODUser, obj models.ODObject) *
 		UserAOsLruCache: ccache.New(ccache.Configure().MaxSize(1000).ItemsToPrune(50)),
 		TypeLruCache:    ccache.New(ccache.Configure().MaxSize(100).ItemsToPrune(5)),
 		EventQueue:      fakeQueue,
+		Conf:            FakeServerSettingsConfiguration(),
 	}
 
 	whitelistedDN := "cn=twl-server-generic2,ou=dae,ou=dia,ou=twl-server-generic2,o=u.s. government,c=us"